Assert properly that the query is not invalidated
